change hook filenames
