+++ hooks/go/070-generate-certs/linstor_certs.go
@@ -1,0 +1,7 @@
+package generatecerts
+
+import "github.com/deckhouse/sds-replicated-volume/hooks/go/certs"
+
+func init() {
+	certs.RegisterLinstorCertsHook()
+}
