--- hooks/go/070-generate-certs/register_all.go
@@ -1,12 +1,0 @@
-package generatecerts
-
-import (
-	"github.com/deckhouse/sds-replicated-volume/hooks/go/certs"
-)
-
-func init() {
-	certs.RegisterLinstorCertsHook()
-	certs.RegisterWebhookCertsHook()
-	certs.RegisterSchedulerExtenderCertHook()
-	certs.RegisterSpaasCertHook()
-}
