+++ hooks/go/070-generate-certs/scheduler_extender_cert.go
@@ -1,0 +1,7 @@
+package generatecerts
+
+import "github.com/deckhouse/sds-replicated-volume/hooks/go/certs"
+
+func init() {
+	certs.RegisterSchedulerExtenderCertHook()
+}
