+++ hooks/go/070-generate-certs/webhook_certs.go
@@ -1,0 +1,7 @@
+package generatecerts
+
+import "github.com/deckhouse/sds-replicated-volume/hooks/go/certs"
+
+func init() {
+	certs.RegisterWebhookCertsHook()
+}
